requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
python-dotenv>=0.19.0
google-generativeai>=0.3.0